--- a/pkg/ingester/ingester_test.go
+++ b/pkg/ingester/ingester_test.go
@@ -10,6 +10,8 @@ import (
 	"testing"
 	"time"
 
+	"github.com/grafana/loki/pkg/distributor"
+
 	"github.com/grafana/dskit/flagext"
 	"github.com/grafana/dskit/services"
 	"github.com/prometheus/common/model"
@@ -784,7 +786,7 @@ func Test_DedupeIngesterParser(t *testing.T) {
 	defer closer()
 
 	for i := 0; i < streamCount; i++ {
-		streams = append(streams, labels.FromStrings("foo", "bar", "bar", fmt.Sprintf("baz%d", i)))
+		streams = append(streams, labels.FromStrings("foo", "bar", "bar", fmt.Sprintf("baz%d", i), distributor.ShardLbName, fmt.Sprint(i%2)))
 	}
 
 	for i := 0; i < requests; i++ {