--- a/pkg/storage/async_store_test.go
+++ b/pkg/storage/async_store_test.go
@@ -66,7 +66,7 @@ func buildMockChunkRef(t *testing.T, num int) []chunk.Chunk {
 	}
 
 	for i := 0; i < num; i++ {
-		chk := newChunk(buildTestStreams(fooLabelsWithName, timeRange{
+		chk := newChunk(buildTestStreams(fooLabelsWithNameAndShard, timeRange{
 			from: now.Add(time.Duration(i) * time.Minute),
 			to:   now.Add(time.Duration(i+1) * time.Minute),
 		}))