… pkg/storage/batch_test.go