As streams are sharded and resharded across different ingesters, the likelihood of duplicate logs being ingested increases dramatically. This removes duplicates at query time by dynamically removing the `__shard_stream__` label  for both log and sample queries so the `stream hash` used by the `MergeEntryIterator` will match when doing comparisons for duplicates.

The `stream hash` is created in either the `store` or `ingester instance` at query time so the labels are removed there. Additionally, the `MergeEntryIterator` has been modified to remove duplicates in a stream rather than just across streams. 

For completeness, `__shard_stream__` has also been removed from Label and Series Queries.