@MasslessParticle 2022-08-30
This change had a minimal affect on benchmarks:
```
benchmark                               old ns/op     new ns/op     delta
BenchmarkSortIterator/merge_sort-8      4186835       4243637       +1.36%
BenchmarkSortIterator/sort-8            1338503       1330357       -0.61%
BenchmarkSortSampleIterator/merge-8     1655672       1619700       -2.17%
BenchmarkSortSampleIterator/sort-8      287801        288213        +0.14%

benchmark                               old allocs     new allocs     delta
BenchmarkSortIterator/merge_sort-8      6              6              +0.00%
BenchmarkSortIterator/sort-8            104            104            +0.00%
BenchmarkSortSampleIterator/merge-8     6              6              +0.00%
BenchmarkSortSampleIterator/sort-8      3              3              +0.00%

benchmark                               old bytes     new bytes     delta
BenchmarkSortIterator/merge_sort-8      11104         11104         +0.00%
BenchmarkSortIterator/sort-8            318808        318809        +0.00%
BenchmarkSortSampleIterator/merge-8     9416          9416          +0.00%
BenchmarkSortSampleIterator/sort-8      1928          1928          +0.00%
```